pyproject.toml
[project]
name = "things-mcp"
version = "0.1.0"
description = "Model Context Protocol server for Things 3 app integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "things-py>=0.0.15",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
things-mcp = "things_mcp.mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/things_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[dependency-groups]
dev = [
    "bandit>=1.8.5",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
    "ty>=0.0.1a11",
]
